Move back into shaders module
