+++ packages/vfx/src/layers/animateScale.ts
@@ -1,5 +1,5 @@
 import { Vector3 } from "three"
-import { createShader } from "../lib/shadermaker"
+import { createShader } from "../shaders"
 
 export default function(fun = "v_progress") {
   const configurator = {
