+++ packages/vfx/src/layers/provideDepthTexture.ts
@@ -1,5 +1,5 @@
 import { DepthTexture } from "three"
-import { createShader } from "../lib/shadermaker"
+import { createShader } from "../shaders"
 
 export default function(depthTexture: DepthTexture) {
   return createShader({
