+++ packages/vfx/src/layers/provideEasingFunctions.ts
@@ -1,6 +1,6 @@
 /* Big thanks to https://github.com/glslify/glsl-easings for the code below! */
 
-import { createShader } from "../lib/shadermaker"
+import { createShader } from "../shaders"
 
 const functions = /*glsl*/ `
   #ifndef HALF_PI
