+++ packages/vfx/src/shaders/generators.ts
